move api functions into their own file
